# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kuchi Backend API",
        "description": "API documentation for Kuchi Backend with JWT Authentication",
        "contact": {
            "email": "admin@kuchi.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/privacy-policy": {
            "get": {
                "tags": [
                    "Privacy Policy"
                ],
                "summary": "Get all Privacy Policy (Public)",
                "operationId": "4ed053eb3ec7774742ef7ae9e16c3364",
                "parameters": [
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of privacy policies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "General Privacy Policy"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "This privacy policy describes how we handle your data..."
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "user_id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/privacy-policy/{id}": {
            "get": {
                "tags": [
                    "Privacy Policy"
                ],
                "summary": "Get single privacy policy by ID (Public)",
                "operationId": "537dd323e05ce7115b6faef888d353dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Privacy Policy ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Privacy policy details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Privacy policy not found"
                    }
                }
            }
        },
        "/api/admin/privacy-policy": {
            "post": {
                "tags": [
                    "Privacy Policy"
                ],
                "summary": "Create new privacy policy (Admin only)",
                "operationId": "9217ed8b81f412d88f91404cb01368aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "General Privacy Policy"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "This privacy policy describes how we collect, use, and protect your personal data..."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Privacy policy created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Privacy policy created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/privacy-policy/{id}": {
            "put": {
                "tags": [
                    "Privacy Policy"
                ],
                "summary": "Update privacy policy (Admin only)",
                "operationId": "6c5a13634455b6ce08b8f8c72a3630af",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Privacy Policy ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Privacy Policy"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated privacy policy content..."
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Privacy policy updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Privacy policy updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Privacy policy not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Privacy Policy"
                ],
                "summary": "Delete privacy policy (Admin only)",
                "operationId": "879b7b61a570be24eb44783030354e2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Privacy Policy ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Privacy policy deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Privacy policy deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Privacy policy not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/privacy-policy/latest/active": {
            "get": {
                "tags": [
                    "Privacy Policy"
                ],
                "summary": "Get latest active privacy policy (Public)",
                "operationId": "097ed80bdad1fd3428f92007c16165e1",
                "responses": {
                    "200": {
                        "description": "Latest active privacy policy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active privacy policy found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}